+++ src/TheatricalPlayersRefactoringKata.Application/Printer/XmlStatementPrinter.cs
@@ -1,11 +1,10 @@
-﻿using System.Collections.Generic;
-using System.Globalization;
-using System.IO;
+﻿using System.Globalization;
 using System.Text;
 using System.Xml;
-using System;
+using TheatricalPlayersRefactoringKata.Application.Calculator;
+using TheatricalPlayersRefactoringKata.Entities;
 
-namespace TheatricalPlayersRefactoringKata
+namespace TheatricalPlayersRefactoringKata.Application.Printer
 {
     public class XmlStatementPrinter : IStatementPrinter
     {
