+++ src/TheatricalPlayersRefactoringKata/Entities/Performance.cs
@@ -1,4 +1,4 @@
-namespace TheatricalPlayersRefactoringKata;
+namespace TheatricalPlayersRefactoringKata.Entities;
 
 public class Performance
 {
@@ -10,8 +10,8 @@ public class Performance
 
     public Performance(string playID, int audience)
     {
-        this._playId = playID;
-        this._audience = audience;
+        _playId = playID;
+        _audience = audience;
     }
 
 }
