+++ src/TheatricalPlayersRefactoringKata/Entities/Play.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace TheatricalPlayersRefactoringKata;
+namespace TheatricalPlayersRefactoringKata.Entities;
 
 public class Play
 {
@@ -12,10 +12,11 @@ public class Play
     public int Lines { get => _lines; set => _lines = value; }
     public string Type { get => _type; set => _type = value; }
 
-    public Play(string name, int lines, string type) {
-        this._name = name;
-        this._lines = lines;
-        this._type = type;
+    public Play(string name, int lines, string type)
+    {
+        _name = name;
+        _lines = lines;
+        _type = type;
     }
 
     public int CalculateBaseAmount()
