+++ tests/TheatricalPlayersRefactoringKata.Tests/StatementPrinterTests.TestStatementExampleLegacy.approved.txt
