+++ tests/TheatricalPlayersRefactoringKata.Tests/StatementPrinterTests.TestTextStatementExample.approved.txt
