+++ tests/TheatricalPlayersRefactoringKata.Tests/StatementPrinterTests.TestXmlStatementExample.approved.txt
